Add external beacon LEDs, systemd unit, fix font-viewer type error
- Add 3 external LEDs (P0.06, P0.08, P1.11) with aircraft anti-collision
  beacon double-flash pattern
- Add systemd service unit for voyager-host deployment on RPi
- Fix u64→u32 cast in font-viewer for Distance::from_unix_timestamp

+++ font-viewer/src/main.rs
@@ -89,7 +89,7 @@ fn run_viewer(bmp: &Bmp<BinaryColor>, bmp_width: i32, scale: u32) {
             clear_display(&mut display).unwrap();
 
             // Calculate real Voyager 1 distance
-            let distance = Distance::from_unix_timestamp(unix_secs);
+            let distance = Distance::from_unix_timestamp(unix_secs as u32);
             let delay = distance.signal_delay();
 
             // Render info section (top half)
